Centralize iframe constants, drop dead test variable
Address ce:review findings:
- Move UNSUPPORTED_IFRAME_SRCS, frame-depth clamping, and the
  ignore-selector resolver into iframe-utils.js so the same constants
  don't drift across SDKs.
- isUnsupportedIframeSrc lowercases the src so JavaScript:/DATA:
  prefixes are caught.
- Drop unused defaultContentCalls counter flagged by lint.
